+++ include/PictureIt/effects/crossfade.h
