+++ include/PictureIt/effects/effects.h
