+++ include/PictureIt/effects/slide.h
