+++ include/PictureIt/spectrum.h
