+++ include/PictureIt/utils.h
